+++ src/services/peer.ts
@@ -2,7 +2,7 @@ import { Socket } from 'socket.io';
 
 import Base from './base';
 import { HEARTBEAT_TIMEOUT } from '../lib/contants';
-import { HandState, PeerData, Role, Tag } from '../types/interfaces';
+import { HandState, PeerData, Role, Tag } from '../types';
 
 class Peer extends Base {
   isRecorder: boolean;
